Move custom lists pydantic model into models package
